% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_Expression_limma.R
\name{differential_Expression_limma}
\alias{differential_Expression_limma}
\title{Preforms Differential Gene Expression Using Limma}
\usage{
differential_Expression_limma(
  gse,
  sampleInfo,
  threshold_genes = 2,
  col1 = "col1",
  col2 = "col2"
)
}
\arguments{
\item{gse}{the gse object from import_GEO}

\item{sampleInfo}{the phenotype data from sample_simplify}

\item{threshold_genes}{the threshold for the number of times a gene needs to be expressed default = 2}

\item{col1}{the name the user wants for column 1 default = "col1"}

\item{col2}{the name the user wants for column 2 default = "col2"}
}
\value{
returns the fit2 object which contains the results of the differential expression for further plotting
}
\description{
Preforms differential gene expression on the gse object using Limma package.
}
\examples{
differential_Expression_limma(gse,sampleInfo, 3, "tumor", "normal")
}
