% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_heatmap.R
\name{GEO_heatmap}
\alias{GEO_heatmap}
\title{Generates a Pheatmap from the GSE Object}
\usage{
GEO_heatmap(Expression_data, sampleInfo = NaN)
}
\arguments{
\item{Expression_data}{the Expression_data from exprs(gse) from GEO_data}

\item{sampleInfo}{optional simplified phenotype data from sample_simplify to annotate the heatmap.}
}
\value{
This produces a heatmap of the gene expression of the clusters. It can either be annotated of unannotated.
}
\description{
This function generates a heatmap from the gs object. If sample_info is provided from sample_simplify then the heatmap will be annotated otherwise it will not be.
}
\examples{
GEO_heatmap(gse), GEO_heatmap(gse,sampleInfo)
}
