% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_GEO.R
\name{import_GEO}
\alias{import_GEO}
\title{Import GEO object from given ID}
\usage{
import_GEO(GEO_id)
}
\arguments{
\item{GEO_id}{The GEO ID you want to collected for analysis}
}
\value{
Returns gse the object which contains all the data from the GEO of your given ID.
}
\description{
Collects the GSE data of a given GEO ID from the gene expression omnibus.
}
\examples{
import_GEO("GSE33126")
}
