% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_plot.R
\name{MA_plot}
\alias{MA_plot}
\title{Generate MA Plot from Limma Differential Expression}
\usage{
MA_plot(
  fit2,
  top_genes_frame,
  xlabel = "Average log-expression",
  ylabel = "Expression log-ratio (this vs that)"
)
}
\arguments{
\item{fit2}{the limma results object from differential_Expression_limma}

\item{top_genes_frame}{the top genes frame from annotate_fit}

\item{xlabel}{the label of x axis default = "Average log-expression"}

\item{ylabel}{the label of y axis default = "Expression log-ratio (this vs that)"}
}
\value{
creates an MA plot of the limma object
}
\description{
A function which generates a MA Plot from the fit2 object
}
\examples{
MA_plot(fit2, top_genes_frame,"Log expression", "cancer vs normal")
}
