% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Creates a Scatter Plot of the PCA Vectors}
\usage{
PCA(Expression_data, sampleInfo)
}
\arguments{
\item{Expression_data}{the Expression_data from exprs(gse) from GEO_data}

\item{sampleInfo}{the phenotype data from sample_simplify}
}
\value{
Returns the scatter plot of the PCA object.
}
\description{
Creates a PCA analysis of the gse object and produces a scatter plot from them.
}
\examples{
PCA(gse,sampleInfo)
}
