% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis_enrich.R
\name{reactome_analysis_enrich}
\alias{reactome_analysis_enrich}
\title{Pathway Enrichment Analysis of Differently Expressed Genes Using ReactomePA}
\usage{
reactome_analysis_enrich(top_genes_frame, p_val_cutoff = 0.05)
}
\arguments{
\item{top_genes_frame}{the top genes data from annotate_fit}

\item{p_val_cutoff}{the p value cut off for determining significance default = 0.05}
}
\value{
returns the enrichPathway results object
}
\description{
Uses the top gene frame to preform overly enriched differently expressed genes.
}
\examples{
reactome_analysis_enrich(top_genes_frame,0.01)
}
