% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Simplify.R
\name{sample_Simplify}
\alias{sample_Simplify}
\title{Simplifies Phenotype Data}
\usage{
sample_Simplify(sampleInfo)
}
\arguments{
\item{sampleInfo}{the Phenotype_data from the gse object out_list from GEO_data.}
}
\value{
returns the same sampleInfo Phenotype_data object but altered to be simplified.
}
\description{
Simplifies Phenotype Data via selecting and renaming the phenotype data from the gse object.
}
\examples{
sample_Simplify(subj)
}
