% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Generate Volcano Plot from Limma Differential Expression}
\usage{
volcano_plot(top_genes_frame, fit2, p_val_cutoff = 0.05)
}
\arguments{
\item{top_genes_frame}{the top genes frame from annotate_fit}

\item{fit2}{the limma results object from differential_Expression_limma}

\item{p_val_cutoff}{the p value cut off for significance default = 0.05}
}
\value{
creates a volcano plot of the Limma data
}
\description{
A function which generates a Volcano Plot from the fit2 object
}
\examples{
volcano_plot(top_genes_frame, fit2,0.01)
}
